Rename and repackage mixin service
