+++ architectury-mixin-remapper-service/src/main/resources/META-INF/services/cpw.mods.modlauncher.api.ITransformationService
@@ -1,1 +1,1 @@
-dev.architectury.loom.forgeruntime.mixin.ForgeLoomMixinRemapperInjectorService
+dev.architectury.mixinremapperservice.ArchitecturyMixinRemapperInjectorService
